Add blog and about pages, and create post collection
This commit introduces two new pages: blog and about, and generates a post collection schema in config.ts. The blog page serves as the main blog post listing, while the about page provides information about the blog. A collection for posts has been defined to facilitate the structuring and displaying of future blog post content.

Furthermore, transition views have been added to the layout component for a smoother UI experience. Another markdown post 'behind-the-scenes-with-our-artists.md' was also added to the content/posts directory. Lastly, a reference path was added to ease the typescript compilation process.