Replace astro.yml with deploy.yml for GitHub Actions workflow
The astro.yml file, used in the GitHub Actions workflow, was deleted and replaced by the deploy.yml file. This change was made because some unnecessary steps in the astro.yml (like installation of pnpm and sharp) were removed to streamline the workflow. The new deploy.yml covers all the necessary steps including repository check out, site build and deployment to GitHub Pages.