Add Sharp installation to GitHub Actions workflow
The Sharp library is added to our Github Actions workflow configuration. This change was necessary because Sharp is a crucial dependency required for image processing in our build process. Its absence caused the builds to fail before. Now, with Sharp installed as part of the workflow, we ensure consistent and successful builds.