Update deploy workflow to use pnpm for dependency management
Updated the '.github/workflows/deploy.yml' file to use pnpm as a package manager. This change includes the installation of pnpm and the modification of the dependencies installation step to use pnpm. This is done to take advantage of pnpm's efficient package management and its ability to improve build times.